Basculer les vérifications dans les controllers
* ♻️ Refactor(npc): Move validation logic in the controller

* ♻️ Refactor(Player): Move validation logic in the controller

* ♻️ Refactor(Character): Move validation logic in the controller

* ♻️ Refactor(groups): Move validation logic in the controller

* ♻️ Refactor(campaign): Move validation logic in the controller

* 🔥 Refactor(campaigns): Remove useless error handling

* ♻️ Refactor(User): Move validation logic into controller

* ♻️ Refactor(auth): Move validation logic into controller

* ♻️ Refactor(auth): Implement pipe for validating MongoDB ObjectId

* ♻️ Refactor(user): Implement pipe for validating MongoDB ObjectId

* ♻️ Refactor(campaigns): Implement pipe for validating MongoDB ObjectId

* ♻️ Refactor(groups): Implement pipe for validating mongoseId

* ♻️ Refactor(character): Implement pipe for validating mongooseId

* ♻️ Refactor(npc): Implement pipe for validating mongooseId

* ♻️ Refactor(player): Implement pipe for validating mongooseId

* ♻️ Refactor(character): Final fix

* Fix import

---------
