+++ backend/src/resources/campaign/campaign.service.ts
@@ -1,10 +1,8 @@
 import {
   BadRequestException,
-  GoneException,
   Injectable,
   InternalServerErrorException,
   Logger,
-  NotFoundException,
 } from '@nestjs/common';
 import { CreateCampaignDto } from '@/resources/campaign/dto/create-campaign.dto';
 import { UpdateCampaignDto } from '@/resources/campaign/dto/update-campaign.dto';
@@ -21,66 +19,15 @@ export class CampaignService {
   constructor(
     @InjectModel(Campaign.name) private campaignModel: Model<CampaignDocument>,
     @InjectModel(Group.name) private groupModel: Model<GroupDocument>,
-  ) {}
+  ) { }
 
   private readonly logger = new Logger(CampaignService.name);
   private readonly SERVICE_NAME = CampaignService.name;
-
-  private async validateGroupRelations(
-    groupIds: string[],
-    type: 'Main' | 'NPC' | 'Archived',
-  ): Promise<void> {
-    if (!groupIds || groupIds.length === 0) return;
-
-    for (const groupId of groupIds) {
-      if (!Types.ObjectId.isValid(groupId)) {
-        throw new BadRequestException(`Invalid ${type} group ID: ${groupId}`);
-      }
-
-      const group = await this.groupModel.findById(groupId).exec();
-
-      if (!group) {
-        throw new NotFoundException(`${type} group not found: ${groupId}`);
-      }
-
-      if (group.deletedAt) {
-        throw new GoneException(`${type} group deleted: ${groupId}`);
-      }
-    }
-  }
-
-  private async validateResource(id: string): Promise<void> {
-    if (!Types.ObjectId.isValid(id)) {
-      const message = `Error while fetching campaign #${id}: Id is not a valid mongoose id`;
-      this.logger.error(message, null, this.SERVICE_NAME);
-      throw new BadRequestException(message);
-    }
-    const campaign = await this.campaignModel.findById(id).exec();
-
-    if (!campaign) {
-      const message = `Campaign #${id} not found`;
-      this.logger.error(message, null, this.SERVICE_NAME);
-      throw new NotFoundException(message);
-    }
-
-    if (campaign.deletedAt) {
-      const message = `Campaign #${id} is gone`;
-      this.logger.error(message, null, this.SERVICE_NAME);
-      throw new GoneException(message);
-    }
-  }
 
   async create(createCampaignDto: CreateCampaignDto, userId: string) {
     try {
       const { groups, ...campaignData } = createCampaignDto;
       const totalGroups = groups.main.concat(groups.npc, groups.archived);
-
-      await this.validateGroupRelations(createCampaignDto.groups.main, 'Main');
-      await this.validateGroupRelations(createCampaignDto.groups.npc, 'NPC');
-      await this.validateGroupRelations(
-        createCampaignDto.groups.archived,
-        'Archived',
-      );
 
       const start: number = Date.now();
       const campaign = await this.campaignModel.create({
@@ -101,13 +48,6 @@ export class CampaignService {
         data: campaign,
       };
     } catch (error) {
-      if (
-        error instanceof BadRequestException ||
-        error instanceof GoneException ||
-        error instanceof NotFoundException
-      ) {
-        throw error;
-      }
       const message = `Error while creating campaign: ${error.message}`;
       this.logger.error(message, null, this.SERVICE_NAME);
       throw new InternalServerErrorException(message);
@@ -194,10 +134,8 @@ export class CampaignService {
     }
   }
 
-  async findOne(id: string) {
+  async findOne(id: Types.ObjectId) {
     try {
-      await this.validateResource(id);
-
       const start: number = Date.now();
       const campaign = await this.campaignModel
         .findById(id)
@@ -214,39 +152,16 @@ export class CampaignService {
         data: campaign,
       };
     } catch (error) {
-      if (
-        error instanceof NotFoundException ||
-        error instanceof GoneException ||
-        error instanceof BadRequestException
-      ) {
-        throw error;
-      }
-
       const message = `Error while fetching campaign #${id}: ${error.message}`;
       this.logger.error(message, null, this.SERVICE_NAME);
       throw new InternalServerErrorException(message);
     }
   }
 
-  async update(id: string, updateCampaignDto: UpdateCampaignDto) {
+  async update(id: Types.ObjectId, updateCampaignDto: UpdateCampaignDto) {
     try {
-      await this.validateResource(id);
-
-      if (updateCampaignDto.groups) {
-        await this.validateGroupRelations(
-          updateCampaignDto.groups.main,
-          'Main',
-        );
-        await this.validateGroupRelations(updateCampaignDto.groups.npc, 'NPC');
-        await this.validateGroupRelations(
-          updateCampaignDto.groups.archived,
-          'Archived',
-        );
-      }
-
       const start = Date.now();
 
-      // Find existing campaign
       const existingCampaign = await this.campaignModel.findById(id);
 
       // Handle label update attempt
@@ -319,9 +234,7 @@ export class CampaignService {
       };
     } catch (error) {
       if (
-        error instanceof BadRequestException ||
-        error instanceof NotFoundException ||
-        error instanceof GoneException
+        error instanceof BadRequestException
       ) {
         throw error;
       }
@@ -331,10 +244,8 @@ export class CampaignService {
     }
   }
 
-  async remove(id: string) {
+  async remove(id: Types.ObjectId) {
     try {
-      await this.validateResource(id);
-
       const start: number = Date.now();
 
       const campaign = await this.campaignModel.findById(id).exec();
@@ -359,14 +270,6 @@ export class CampaignService {
         data: campaign,
       };
     } catch (error) {
-      if (
-        error instanceof BadRequestException ||
-        error instanceof NotFoundException ||
-        error instanceof GoneException
-      ) {
-        throw error;
-      }
-
       const message = `Error while deleting campaign #${id}: ${error.message}`;
       this.logger.error(message, null, this.SERVICE_NAME);
       throw new InternalServerErrorException(message);
