+++ backend/src/resources/user/user.service.ts
@@ -1,7 +1,5 @@
 import {
-  BadRequestException,
   ConflictException,
-  GoneException,
   Injectable,
   InternalServerErrorException,
   Logger,
@@ -12,18 +10,13 @@ import { UpdateUserDto } from '@/resources/user/dto/update-user.dto';
 import { InjectModel } from '@nestjs/mongoose';
 import { User, UserDocument } from '@/resources/user/schemas/user.schema';
 import { Model, Types } from 'mongoose';
-import {
-  Campaign,
-  CampaignDocument,
-} from '@/resources/campaign/schemas/campaign.schema';
 import * as bcrypt from 'bcrypt';
 
 @Injectable()
 export class UserService {
   constructor(
     @InjectModel(User.name) private userModel: Model<UserDocument>,
-    @InjectModel(Campaign.name) private campaignModel: Model<CampaignDocument>,
-  ) {}
+  ) { }
 
   private readonly SERVICE_NAME = UserService.name;
   private readonly logger = new Logger(this.SERVICE_NAME);
@@ -40,33 +33,6 @@ export class UserService {
         const message = `User with email ${userData.email} already exists`;
         this.logger.error(message, null, this.SERVICE_NAME);
         throw new ConflictException(message);
-      }
-
-      const campaignCheckPromises = campaigns.map((campaignId) =>
-        this.campaignModel.findById(campaignId).exec(),
-      );
-      const campaignCheckResults = await Promise.all(campaignCheckPromises);
-      const invalidCampaigns = campaignCheckResults.filter(
-        (campaign) => !campaign,
-      );
-      if (invalidCampaigns.length > 0) {
-        const invalidCampaignIds = campaigns.filter(
-          (_, index) => !campaignCheckResults[index],
-        );
-        const message = `Invalid campaign IDs: ${invalidCampaignIds.join(', ')}`;
-        this.logger.error(message, null, this.SERVICE_NAME);
-        throw new BadRequestException(message);
-      }
-      const goneCampaigns = campaignCheckResults.filter(
-        (campaign) => campaign.deletedAt !== null,
-      );
-      if (goneCampaigns.length > 0) {
-        const goneCampaignIds = campaigns.filter(
-          (_, index) => campaignCheckResults[index].deletedAt !== null,
-        );
-        const message = `Gone campaign IDs: ${goneCampaignIds.join(', ')}`;
-        this.logger.error(message, null, this.SERVICE_NAME);
-        throw new GoneException(message);
       }
 
       // Hash the password
@@ -89,8 +55,6 @@ export class UserService {
       };
     } catch (error) {
       if (
-        error instanceof BadRequestException ||
-        error instanceof GoneException ||
         error instanceof ConflictException
       ) {
         throw error;
@@ -153,29 +117,11 @@ export class UserService {
     }
   }
 
-  async findOne(id: string) {
+  async findOne(id: Types.ObjectId) {
     try {
-      if (!Types.ObjectId.isValid(id)) {
-        const message = `Error while fetching campaign #${id}: Id is not a valid mongoose id`;
-        this.logger.error(message, null, this.SERVICE_NAME);
-        throw new BadRequestException(message);
-      }
-
       const start: number = Date.now();
       const user = await this.userModel.findById(id).exec();
       const end: number = Date.now();
-
-      if (!user) {
-        const message = `User #${id} not found`;
-        this.logger.error(message, null, this.SERVICE_NAME);
-        throw new NotFoundException(message);
-      }
-
-      if (user.deletedAt) {
-        const message = `User #${id} is gone`;
-        this.logger.error(message, null, this.SERVICE_NAME);
-        throw new GoneException(message);
-      }
 
       const message = `User #${id} found in ${end - start}ms`;
       this.logger.verbose(message, this.SERVICE_NAME);
@@ -184,13 +130,6 @@ export class UserService {
         data: user,
       };
     } catch (error) {
-      if (
-        error instanceof NotFoundException ||
-        error instanceof GoneException ||
-        error instanceof BadRequestException
-      ) {
-        throw error;
-      }
       const message = `Error while getting user #${id}: ${error.message}`;
       this.logger.error(message, null, this.SERVICE_NAME);
       throw new InternalServerErrorException(message);
@@ -226,62 +165,13 @@ export class UserService {
     }
   }
 
-  async update(id: string, updateUserDto: UpdateUserDto) {
+  async update(id: Types.ObjectId, updateUserDto: UpdateUserDto) {
     try {
-      if (!Types.ObjectId.isValid(id)) {
-        const message = `Error while updating user #${id}: Id is not a valid mongoose id`;
-        this.logger.error(message, null, this.SERVICE_NAME);
-        throw new BadRequestException(message);
-      }
-      let user = await this.userModel.findById(id);
-
-      if (!user) {
-        const message = `User #${id} not found`;
-        this.logger.error(message, null, this.SERVICE_NAME);
-        throw new NotFoundException(message);
-      }
-
-      if (user.deletedAt) {
-        const message = `User #${id} deleted`;
-        this.logger.error(message, null, this.SERVICE_NAME);
-        throw new GoneException(message);
-      }
-
-      const { campaigns = [], ...userData } = updateUserDto;
-
-      const campaignCheckPromises = campaigns.map((campaignId) =>
-        this.campaignModel.findById(campaignId).exec(),
-      );
-      const campaignCheckResults = await Promise.all(campaignCheckPromises);
-      const invalidCampaigns = campaignCheckResults.filter(
-        (campaign) => !campaign,
-      );
-      if (invalidCampaigns.length > 0) {
-        const invalidCampaignIds = campaigns.filter(
-          (_, index) => !campaignCheckResults[index],
-        );
-        const message = `Invalid campaign IDs: ${invalidCampaignIds.join(', ')}`;
-        this.logger.error(message, null, this.SERVICE_NAME);
-        throw new BadRequestException(message);
-      }
-
-      const goneCampaigns = campaignCheckResults.filter(
-        (campaign) => campaign.deletedAt !== null,
-      );
-      if (goneCampaigns.length > 0) {
-        const goneCampaignIds = campaigns.filter(
-          (_, index) => campaignCheckResults[index].deletedAt !== null,
-        );
-        const message = `Gone campaign IDs: ${goneCampaignIds.join(', ')}`;
-        this.logger.error(message, null, this.SERVICE_NAME);
-        throw new GoneException(message);
-      }
-
       const start: number = Date.now();
       const userUpdate = await this.userModel
         .updateOne({ _id: id }, updateUserDto)
         .exec();
-      user = await this.userModel.findById(id).populate('campaigns').exec();
+      const user = await this.userModel.findById(id).populate('campaigns').exec();
       const end: number = Date.now();
 
       if (!user || userUpdate.modifiedCount === 0) {
@@ -298,9 +188,7 @@ export class UserService {
       };
     } catch (error) {
       if (
-        error instanceof NotFoundException ||
-        error instanceof GoneException ||
-        error instanceof BadRequestException
+        error instanceof NotFoundException
       ) {
         throw error;
       }
@@ -310,29 +198,10 @@ export class UserService {
     }
   }
 
-  async remove(id: string) {
+  async remove(id: Types.ObjectId) {
     try {
-      if (!Types.ObjectId.isValid(id)) {
-        const message = `Error while deleting user #${id}: Id is not a valid mongoose id`;
-        this.logger.error(message, null, this.SERVICE_NAME);
-        throw new BadRequestException(message);
-      }
-
       const start: number = Date.now();
-
       const user = await this.userModel.findById(id).exec();
-
-      if (!user) {
-        const message = `User #${id} not found`;
-        this.logger.error(message, null, this.SERVICE_NAME);
-        throw new NotFoundException(message);
-      }
-
-      if (user.deletedAt) {
-        const message = `User #${id} already deleted`;
-        this.logger.error(message, null, this.SERVICE_NAME);
-        throw new GoneException(message);
-      }
 
       user.deletedAt = new Date();
       await user.save();
@@ -346,14 +215,6 @@ export class UserService {
         data: user,
       };
     } catch (error) {
-      if (
-        error instanceof BadRequestException ||
-        error instanceof NotFoundException ||
-        error instanceof GoneException
-      ) {
-        throw error;
-      }
-
       const message = `Error while deleting character #${id}: ${error.message}`;
       this.logger.error(message, null, this.SERVICE_NAME);
       throw new InternalServerErrorException(message);
